#1 refactor: 메서드 기능별로 class 구분하기
