+++ src/main/java/kiosk/CheckException.java
@@ -1,0 +1,8 @@
+package kiosk;
+
+public class CheckException {
+	public static void checkInputRange(int selectInt){
+		if(selectInt < FunctionAboutFirstInput.MinInputNum || FunctionAboutFirstInput.MaxInputNum < selectInt) throw new IllegalArgumentException("잘못된 메뉴를 선택하셨습니다.");
+	}
+
+}
