+++ src/main/java/kiosk/ContactByUser.java
@@ -1,0 +1,10 @@
+package kiosk;
+
+import java.util.Scanner;
+
+public class ContactByUser {
+	public static int getIntByUser(){
+		Scanner scanner = new Scanner(System.in);
+		return scanner.nextInt();
+	}
+}
