+++ src/main/java/kiosk/Item.java
@@ -1,0 +1,14 @@
+package kiosk;
+
+public class Item {
+	String itemName;
+	int price;
+	int count;
+
+	Item(String itemName, int price){
+		this.itemName = itemName;
+		this.price = price;
+		this.count = 1;
+
+	}
+}
