--- src/main/java/kiosk/Print.java
@@ -1,14 +1,0 @@
-package kiosk;
-
-public class Print {
-	public void printHome(){
-		System.out.println("=====홈=====");
-		System.out.println("1. 햄버거");
-		System.out.println("2. 사이드");
-		System.out.println("3. 음료");
-		System.out.println("4. 장바구니");
-		System.out.println("5. 종료\n");
-		System.out.print("메뉴선택:");
-	}
-
-}
